requirements.txt
langchain==0.3.25
langchain-openai==0.3.21
langchain-community==0.3.24
openai==1.84.0
faiss-cpu==1.11.0
tiktoken==0.9.0
pypdf==5.6.0
python-dotenv==1.1.0
streamlit==1.45.1
numpy==2.3.0
pandas==2.3.0 